+++ protocols/relay/CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.1.0 [unreleased]
+# 0.2.0 [unreleased]
+
+- Update `libp2p-swarm`.
+
+# 0.1.0 [2021-03-17]
 
 - First release supporting all major features of the circuit relay v1
   specification. [PR 1838](https://github.com/libp2p/rust-libp2p/pull/1838).