+++ protocols/request-response/CHANGELOG.md
@@ -1,9 +1,15 @@
-# 0.10.0 [unreleased]
+# 0.11.0 [unreleased]
+
+- Update `libp2p-swarm`.
+
+# 0.10.0 [2021-03-17]
 
 - Update `libp2p-swarm`.
 
 - Close stream even when no response has been sent.
   [PR 1987](https://github.com/libp2p/rust-libp2p/pull/1987).
+
+- Update dependencies.
 
 # 0.9.1 [2021-02-15]